+++ .yarnrc.yml
@@ -71,7 +71,7 @@ packageExtensions:
   "@web3auth/base-plugin@2.x":
     peerDependencies:
       "@babel/runtime": "*"
-  demo-express-proxy@*:
+  demo-express@*:
     peerDependencies:
       "@express": "*"
   demo-nextjs@*:
