+++ demos/express/.env.example
@@ -1,4 +1,4 @@
-PORT = 4000;
+PORT = 4000
 
 # Your Moralis Api key, that can be found in the dashboard. Keep this secret!
 MORALIS_API_KEY = ''
