+++ demos/express/package.json
@@ -1,17 +1,17 @@
 {
-  "name": "demo-express-proxy",
+  "name": "demo-express",
   "version": "1.0.0",
   "main": "dist/index.js",
   "private": true,
   "dependencies": {
     "@moralisweb3/common-core": "^2.11.0",
+    "@moralisweb3/express": "^2.11.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "envalid": "7.3.1",
-    "express": "^4.18.2",
     "express-rate-limit": "^6.7.0",
-    "moralis": "^2.11.0",
-    "redis": "^4.2.0"
+    "express": "^4.18.2",
+    "moralis": "^2.11.0"
   },
   "devDependencies": {
     "@moralisweb3/eslint-config": "^1.0.1",
