+++ demos/express/src/index.ts
@@ -3,7 +3,6 @@ import express from 'express';
 import cors from 'cors';
 import config from './config';
 import { apiRouter } from './apiRouter';
-import { errorHandler } from './middlewares/errorHandler';
 
 const app = express();
 
@@ -17,11 +16,10 @@ app.use(express.json());
 app.use(cors());
 
 app.use('/api', apiRouter);
-app.use(errorHandler);
 
 app.use(express.static('public'));
 
 app.listen(config.PORT, () => {
   // eslint-disable-next-line no-console
-  console.log(`'Moralis Demo Express proxy' is running on port ${config.PORT}`);
+  console.log(`'Moralis Demo Express' is running on port ${config.PORT}`);
 });
